backend/operations-center/requirements.txt
fastapi
uvicorn
httpx
supabase
python-dotenv
pydantic
pydantic-settings
